% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanNames.R
\name{CleanNames}
\alias{CleanNames}
\title{CleanNames}
\usage{
CleanNames(
  species_list,
  fasta_file,
  output_file = NULL,
  remove_unmatched = FALSE
)
}
\arguments{
\item{species_list}{csv file containing a "species" column}

\item{fasta_file}{FASTA file to be cleaned}

\item{output_file}{output file path / name for the cleaned FASTA}

\item{remove_unmatched}{Logical. If TRUE, removes sequences that don't match any species in the list.}
}
\value{
FASTA file

FASTA
}
\description{
Cleans FASTA headers by matching species names to a reference list and renaming them in a standardized format.
}
\examples{
\dontrun{
CleanNames(
species_list = "./path.csv",
fasta_file = "./path.fasta",
output_file = "./path.fasta",
remove_unmatched = FALSE
)
}
}
