% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterSpecs.R
\name{FilterSpecs}
\alias{FilterSpecs}
\title{FilterSpecs}
\usage{
FilterSpecs(
  fasta_file,
  species_list,
  min_hits = 1,
  output_type = c("unmatched", "filtered"),
  output_file = "species_output.csv",
  output_fasta = NULL
)
}
\arguments{
\item{fasta_file}{FASTA file to be filtered}

\item{species_list}{csv file containing a "species" column}

\item{min_hits}{minimum number of matches required for a species to be retained}

\item{output_type}{Either "unmatched" (to return species with 0 hits) or "filtered" (to return species meeting hit threshold)}

\item{output_file}{output file path / name for resulting species csv}

\item{output_fasta}{output FASTA path / name for filtered FASTA}
}
\value{
CSV and FASTA
}
\description{
Filters a species list based on sequence matches in FASTA file. Optionally, returns species with no matches ("unmatched")
or those meeting a minimum hit count ("filtered"). Can also write a filtered FASTA containing only matched species.
}
\examples{
\dontrun{
FilterSpecs(
fasta_file = "./path.fasta",
species_list = "./path.csv",
min_hits = 3,
output_type = "filtered",
output_file = "./path.csv",
output_fasta = "./path.fasta"
)
}
}
