% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenSearch.R
\name{GenSearch}
\alias{GenSearch}
\title{GenSearch}
\usage{
GenSearch(
  species_list,
  gene_list,
  output_file,
  retmax_value = 10,
  desired_length = 18000,
  search_field = "GENE"
)
}
\arguments{
\item{species_list}{csv file containing a "species" column}

\item{gene_list}{csv file containing "gene" column character vector of genes to search}

\item{output_file}{output file path / name for FASTA}

\item{retmax_value}{max number of sequences pulled per query}

\item{desired_length}{max length of sequence}

\item{search_field}{GenBank field to search (e.g., "GENE", "ALL")}
}
\value{
FASTA
}
\description{
Searches GenBank for nucleotide sequences based on a list of species and genes, applies optional
length filtered, and writes resulting sequences to a FASTA file.
}
\examples{
\dontrun{
GenSearch(
species_list = "./path.csv",
gene_list = "./path.csv",
output = "./path.fasta",
retmax = 10,
length = 18000,
search = "GENE"
)
}
}
