% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqSmash.R
\name{SeqSmash}
\alias{SeqSmash}
\title{SeqSmash}
\usage{
SeqSmash(
  species_list,
  fasta_file,
  reference_file,
  output_file,
  replace_gaps = FALSE,
  gap_symbol = "?"
)
}
\arguments{
\item{species_list}{csv file containing a "species" column}

\item{fasta_file}{FASTA files with gene sequences to be aligned}

\item{reference_file}{FASTA file of reference genome}

\item{output_file}{output file path / name for FASTA}

\item{replace_gaps}{Logical. whether to replace "-" gaps with custom symbol}

\item{gap_symbol}{replacement character for gaps ("?" or "N")}
}
\value{
FASTA
}
\description{
Aligns genes to reference genome and creates consensus sequence per species.
}
\examples{
\dontrun{
SeqSmash(
species_list = "./path.csv",
fasta_file = "./path.fasta",
reference_file = "./path.fasta",
outputfile = "./path.fasta",
replace_gaps = TRUE,
gap_symbol = "?"
)
}
}
